+++ CS4N.EnergyHistory/Contracts/Models/Station.cs
@@ -4,7 +4,7 @@
   {
     public int Id { get; set; }
     public string Name { get; set; } = "";
-    public double MaxPowerPeak { get; set; }
+    public int MaxPowerPeak { get; set; }
     //public PowerUnitType PowerUnit { get; set; }
     public List<Inverter> Inverters { get; set; } = new List<Inverter>();
     public List<Module> Modules { get; set; } = new List<Module>();
